refactor: make Number class to none inner class
